Projects/Final_Project_DinaBB/requirements.txt
fitbit==0.3.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
pandas==2.1.4
scikit-learn==1.4.2
joblib==1.3.2
numpy==1.26.4
flask==3.0.3
